Add complete stock portfolio manager with React frontend
Full-stack Indian stock portfolio + mutual funds tracking app:
- Spring Boot 3.5 / Java 21 backend with PostgreSQL (Flyway V1-V14)
- React 19 / Vite / Tailwind CSS frontend with 6 pages
- Groww API integration (holdings sync, order sync, account details)
- Yahoo Finance market data with 5-min cache
- AMFI NAV integration for mutual funds
- Technical analysis engine (SMA/RSI/52-week signals)
- Intraday (MIS) vs delivery (CNC) trade type tracking
- Simplified P&L using Groww clear_cash as source of truth
- Sortable/filterable tables, color-coded P&L, sticky headers
- Data import scripts for Groww Excel reports
- PDF upload for trade reports

+++ src/main/resources/db/migration/V2__create_holding_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE holding (
 
     stock_id BIGINT NOT NULL,
 
-    quantity NUMERIC(20,4) NOT NULL,
+    quantity Integer NOT NULL,
 
     average_buy_price NUMERIC(20,4) NOT NULL,
